moved all glfw specific code in a separate package and abstracted it away
