+++ kool-core/src/desktopMain/kotlin/de/fabmax/kool/pipeline/backend/vk/RenderBackendVk.kt
@@ -8,13 +8,12 @@ import de.fabmax.kool.pipeline.*
 import de.fabmax.kool.pipeline.backend.*
 import de.fabmax.kool.pipeline.backend.gl.pxSize
 import de.fabmax.kool.pipeline.backend.stats.BackendStats
+import de.fabmax.kool.platform.ClientApi
+import de.fabmax.kool.platform.KoolWindowJvm
 import de.fabmax.kool.platform.Lwjgl3Context
-import de.fabmax.kool.platform.glfw.GlfwWindow
 import de.fabmax.kool.scene.Scene
 import de.fabmax.kool.util.*
 import kotlinx.coroutines.CompletableDeferred
-import org.lwjgl.glfw.GLFW
-import org.lwjgl.glfw.GLFWVulkan
 import org.lwjgl.vulkan.VK10.*
 import org.lwjgl.vulkan.VkCommandBuffer
 import java.nio.ByteBuffer
@@ -26,7 +25,7 @@ class RenderBackendVk(val ctx: Lwjgl3Context) : RenderBackendJvm {
     override val apiName: String
     override val deviceName: String
 
-    override val window: GlfwWindow
+    override val window: KoolWindowJvm
 
     override val deviceCoordinates: DeviceCoordinates = DeviceCoordinates.VULKAN
     override val features: BackendFeatures
@@ -58,11 +57,7 @@ class RenderBackendVk(val ctx: Lwjgl3Context) : RenderBackendJvm {
     private val emptyScene = Scene("empty-scene")
 
     init {
-        // tell GLFW to not initialize default OpenGL API before we create the window
-        check(GLFWVulkan.glfwVulkanSupported()) { "Cannot find a compatible Vulkan installable client driver (ICD)" }
-        GLFW.glfwWindowHint(GLFW.GLFW_CLIENT_API, GLFW.GLFW_NO_API)
-
-        window = GlfwWindow(ctx)
+        window = ctx.windowSubsystem.createWindow(ClientApi.UNMANAGED, ctx)
         window.setFullscreen(KoolSystem.configJvm.isFullscreen)
         instance = Instance(this, KoolSystem.configJvm.windowTitle)
         surface = Surface(this)
