+++ kool-core/src/desktopMain/kotlin/de/fabmax/kool/pipeline/backend/vk/Surface.kt
@@ -2,35 +2,17 @@ package de.fabmax.kool.pipeline.backend.vk
 
 import de.fabmax.kool.util.BaseReleasable
 import de.fabmax.kool.util.logD
-import de.fabmax.kool.util.memStack
-import de.fabmax.kool.util.releaseWith
-import org.lwjgl.glfw.GLFW.glfwDestroyWindow
-import org.lwjgl.glfw.GLFW.glfwTerminate
-import org.lwjgl.glfw.GLFWVulkan
-import org.lwjgl.vulkan.KHRSurface
 
 class Surface(val backend: RenderBackendVk) : BaseReleasable() {
-    var surfaceHandle = 0L
-        private set
+    val surfaceHandle = backend.window.createVulkanSurface(backend.instance.vkInstance)
 
     init {
-        memStack {
-            val lp = mallocLong(1)
-            vkCheck(GLFWVulkan.glfwCreateWindowSurface(backend.instance.vkInstance, backend.window.windowHandle, null, lp))
-            surfaceHandle = lp[0]
-        }
-        releaseWith(backend.instance)
         logD { "Created surface" }
     }
 
     override fun doRelease() {
         ReleaseQueue.enqueue {
-            KHRSurface.vkDestroySurfaceKHR(backend.instance.vkInstance, surfaceHandle, null)
-            logD { "Destroyed surface" }
-
-            glfwDestroyWindow(backend.window.windowHandle)
-            glfwTerminate()
-            logD { "Destroyed GLFW window" }
+            backend.window.destroyVulkanSurface(surfaceHandle, backend.instance.vkInstance)
         }
     }
 }
